Move SerialGEMV implementation from KokkosBatched to KokkosBlas
