+++ src/batched/KokkosBatched_Util.hpp
@@ -21,6 +21,7 @@
 
 #include "KokkosKernels_config.h"
 #include "KokkosKernels_SimpleUtils.hpp"
+#include "KokkosBlas_util.hpp"
 
 // TPL macros
 #if defined(KOKKOSKERNELS_ENABLE_TPL_MKL)
@@ -217,11 +218,8 @@ struct AVX {
 };
 
 //////// Tags for BLAS ////////
-struct Trans {
-  struct Transpose {};
-  struct NoTranspose {};
-  struct ConjTranspose {};
-};
+
+using KokkosBlas::Trans;
 
 struct Side {
   struct Left {};
@@ -351,100 +349,7 @@ struct algo_level2_blocked_mb_impl<Kokkos::Experimental::SYCLDeviceUSMSpace> {
 
 #endif
 
-struct Algo {
-  struct Level3 {
-    struct Unblocked {
-      static const char *name() { return "Unblocked"; }
-    };
-    struct Blocked {
-      static const char *name() { return "Blocked"; }
-      // TODO:: for now harwire the blocksizes; this should reflect
-      // regieter blocking (not about team parallelism).
-      // this mb should vary according to
-      // - team policy (smaller) or range policy (bigger)
-      // - space (gpu vs host)
-      // - blocksize input (blk <= 4 mb = 2, otherwise mb = 4), etc.
-#if defined(KOKKOS_IF_ON_HOST)
-      static constexpr KOKKOS_FUNCTION int mb() {
-        KOKKOS_IF_ON_HOST((return 4;))
-        KOKKOS_IF_ON_DEVICE((return 2;))
-      }
-
-#else  // FIXME remove when requiring minimum version of Kokkos 3.6
-      static constexpr KOKKOS_FUNCTION int mb() {
-        return algo_level3_blocked_mb_impl<
-            Kokkos::Impl::ActiveExecutionMemorySpace>::value;
-      }
-
-#endif
-    };
-    struct MKL {
-      static const char *name() { return "MKL"; }
-    };
-    struct CompactMKL {
-      static const char *name() { return "CompactMKL"; }
-    };
-
-    // When this is first developed, unblocked algorithm is a naive
-    // implementation and blocked algorithm uses register blocking variant of
-    // algorithm (manual unrolling). This distinction is almost meaningless and
-    // it just adds more complications. Eventually, the blocked version will be
-    // removed and we only use the default algorithm. For testing and
-    // development purpose, we still leave algorithm tag in the template
-    // arguments.
-    using Default = Unblocked;
-  };
-
-  using Gemm      = Level3;
-  using Trsm      = Level3;
-  using Trmm      = Level3;
-  using Trtri     = Level3;
-  using LU        = Level3;
-  using InverseLU = Level3;
-  using SolveLU   = Level3;
-  using QR        = Level3;
-  using UTV       = Level3;
-
-  struct Level2 {
-    struct Unblocked {};
-    struct Blocked {
-      // TODO:: for now harwire the blocksizes; this should reflect
-      // regieter blocking (not about team parallelism).
-      // this mb should vary according to
-      // - team policy (smaller) or range policy (bigger)
-      // - space (cuda vs host)
-      // - blocksize input (blk <= 4 mb = 2, otherwise mb = 4), etc.
-#if defined(KOKKOS_IF_ON_HOST)
-      static constexpr KOKKOS_FUNCTION int mb() {
-        KOKKOS_IF_ON_HOST((return 4;))
-        KOKKOS_IF_ON_DEVICE((return 1;))
-      }
-
-#else  // FIXME remove when requiring minimum version of Kokkos 3.6
-      static constexpr KOKKOS_FUNCTION int mb() {
-        return algo_level2_blocked_mb_impl<
-            Kokkos::Impl::ActiveExecutionMemorySpace>::value;
-      }
-
-#endif
-    };
-    struct MKL {};
-    struct CompactMKL {};
-
-    // When this is first developed, unblocked algorithm is a naive
-    // implementation and blocked algorithm uses register blocking variant of
-    // algorithm (manual unrolling). This distinction is almost meaningless and
-    // it just adds more complications. Eventually, the blocked version will be
-    // removed and we only use the default algorithm. For testing and
-    // development purpose, we still leave algorithm tag in the template
-    // arguments.
-    using Default = Unblocked;
-  };
-
-  using Gemv   = Level2;
-  using Trsv   = Level2;
-  using ApplyQ = Level2;
-};
+using KokkosBlas::Algo;
 
 struct Util {
   template <typename ValueType>
