+++ src/batched/dense/impl/KokkosBatched_Gemv_TeamVector_Internal.hpp
@@ -7,7 +7,7 @@
 
 #include "KokkosBlas1_set_impl.hpp"
 #include "KokkosBlas1_team_scal_impl.hpp"
-#include "KokkosBatched_InnerMultipleDotProduct_Serial_Impl.hpp"
+#include "KokkosBlas2_serial_gemv_inner_multiple_dot.hpp"
 
 namespace KokkosBatched {
