+++ src/sparse/impl/KokkosSparse_spmv_bsrmatrix_impl.hpp
@@ -538,7 +538,7 @@ struct BsrMatrixSpMVTensorCoreDispatcher {
 //
 
 #include "KokkosBlas.hpp"
-#include "KokkosBatched_Gemv_Serial_Internal.hpp"
+#include "KokkosBlas2_serial_gemv_internal.hpp"
 #include "KokkosBatched_Gemv_TeamVector_Internal.hpp"
 #include "KokkosBatched_Gemm_Serial_Internal.hpp"
 #include "KokkosBatched_Gemm_TeamVector_Internal.hpp"
@@ -618,7 +618,7 @@ struct BSR_GEMV_Functor {
       for (ordinal_type ic = 0; ic < count; ++ic) {
         const auto Aview  = row.block(ic);
         const auto xstart = row.block_colidx(ic) * block_dim;
-        KokkosBatched::SerialGemvInternal<KokkosBatched::Algo::Gemv::Blocked>::
+        KokkosBlas::Impl::SerialGemvInternal<KokkosBlas::Algo::Gemv::Blocked>::
             invoke<value_type, value_type>(
                 block_dim, block_dim, alpha, Aview.data(), block_dim, 1,
                 &m_x(xstart), static_cast<int>(m_x.stride_0()), beta1,
@@ -914,9 +914,9 @@ struct BSR_GEMV_Transpose_Functor {
         const auto ystart = row.block_colidx(ic) * block_dim;
         for (ordinal_type jj = 0; jj < block_dim; ++jj) {
           value_type t(0);
-          KokkosBatched::SerialGemvInternal<
-              KokkosBatched::Algo::Gemv::Blocked>::invoke<value_type,
-                                                          value_type>(
+          KokkosBlas::Impl::SerialGemvInternal<
+              KokkosBlas::Algo::Gemv::Blocked>::invoke<value_type,
+                                                       value_type>(
               1, block_dim, alpha1, Aview.data() + jj, Aview.stride_1(),
               Aview.stride_0(), xview.data(), xview.stride_0(), beta1, &t, 1);
           t *= alpha;
@@ -1561,9 +1561,9 @@ struct BSR_GEMM_Transpose_Functor {
         for (ordinal_type jr = 0; jr < num_rhs; ++jr) {
           for (ordinal_type jj = 0; jj < block_dim; ++jj) {
             value_type t(0);
-            KokkosBatched::SerialGemvInternal<
-                KokkosBatched::Algo::Gemv::Blocked>::invoke<value_type,
-                                                            value_type>(
+            KokkosBlas::Impl::SerialGemvInternal<
+                KokkosBlas::Algo::Gemv::Blocked>::invoke<value_type,
+                                                         value_type>(
                 1, block_dim, alpha1, Aview.data() + jj, Aview.stride_1(),
                 Aview.stride_0(), xview.data() + jr * ldx, xview.stride_0(),
                 beta1, &t, 1);
