+++ unit_test/batched/dense/Test_Batched_Dense.hpp
@@ -13,9 +13,6 @@
 #include "Test_Batched_BatchedGemm.hpp"
 #include "Test_Batched_BatchedGemm_Real.hpp"
 #include "Test_Batched_BatchedGemm_Complex.hpp"
-#include "Test_Batched_SerialGemv.hpp"
-#include "Test_Batched_SerialGemv_Real.hpp"
-#include "Test_Batched_SerialGemv_Complex.hpp"
 #include "Test_Batched_SerialGesv.hpp"
 #include "Test_Batched_SerialGesv_Real.hpp"
 #include "Test_Batched_SerialInverseLU.hpp"
